Make ReadableCharList immutable
Updated the ReadableCharList to construct all the constants in the
static initializer block AND converted them all to immutable via
Collections.unmodifiableList.

I also fixed a range issue.  Prior to this change the '~' character,
unicode 0x7E was not being included. This was due to the
inclusive/exclusivity of IntStream.range.  This needed to be changed to
IntStream.rangeClosed so that the params were inclusive/inclusive.

Addresses Bekreth/jetedge#111