Allow ConstantValueLimiter to accept a Supplier
Altered the ConstantValueLimiter to use a Function<Random,T> at its
core.  This allows the construction of a ConstantValueLimiter to be
accomplished using one of three types.

1. a constant value.
2. a Supplier<T>
3. a Function<Random,T>

This opens up the usage a little bit.  The original issue asked for the
inclusion of a Supplier constructor argument but in order to facilitate
stable random generation it seemed like a good idea to allow clients to
have access to the random value used, hence the addition of
Function<Random,T>.

I also added a unit test for the ConstantValueLimiter that should test
all cases.

*NOTE - the ConstantValueLimiter will throw an IllegalArgumentException
(IAE) if the supplied Supplier or Function are null.

Addresses Bekreth/jetedge#106