Add checkstyle to the build.
Added the checkstyle plugin to the 'validate' phase.  However, it only
runs the 'checkstyle' target (not 'check') and so it doesn't fail the
build on errors.  This seems the most prudent place to start as there
are > 1400 errors with the default configuration.

I figure that follow up PRs could actually fix the checkstyle issues
and/or add proper exclusions and further configuration to checkstyle
itself.

I added the checkstyle-suppressions.xml with some sample patterns to
show how the checkstyle fixups could be 'layered in' instead of a
massive PR.

I also added checkstyle to the reporting section of the POM so it shows
after 'mvn site' is run.

Addresses Bekreth/jetedge#118