Make ReadableCharList Immutable (#121)
* Add tests for ReadableCharList

* Make ReadableCharList immutable

Updated the ReadableCharList to construct all the constants in the
static initializer block AND converted them all to immutable via
Collections.unmodifiableList.

I also fixed a range issue.  Prior to this change the '~' character,
unicode 0x7E was not being included. This was due to the
inclusive/exclusivity of IntStream.range.  This needed to be changed to
IntStream.rangeClosed so that the params were inclusive/inclusive.

Addresses Bekreth/jetedge#111

* Fixed the PojoGeneratorIntegrationTest

The PojoGeneratorIntegrationTest was created when the
ReadableCharacterList's full range was 32-125 (prior to 9a711ac).
Once the range was fixed to be 32-126 the test started failing due to
the character range change.  So, the 'expected' values had to be
updated.