Add tooling to keep track of code coverage  (#124)
* Add jacoco coverage plugin to POM

Add the jacoco plugin so that it runs during 'mvn test'.  It will also
generate a coverage report when 'mvn site' is run provided 'mvn test'
has been run and target/jacoco.exec exists.

Addresses Bekreth/jetedge#86

* Fix FieldPojoAnalyzerTest when run with jacoco

When Jacoco runs and gathers coverage it creates a $jacocoData on
classes it's testing.  This failed the FieldPojoAnalyzerTest because it
didn't know to ignore this field.  Now it does.

* Move surefire.version up into pom.xml properties

* Make the field accessible for testing

* Ignore $jacocoData in reflection code in test

Jacoco adds $jacocoData to classes and the Immutability test was
stumbling on it.  I added an ignore.  I don't like this sort of thing
but coverage is more important than my sensibilities or some other more
complicated solution IMO.